# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lwang Black API",
    "description": "E-commerce backend API for Lwang Black specialty coffee. Supports multi-currency payments (Stripe, PayPal, Khalti, eSewa, Nabil Bank), global logistics (Shippo, DHL, FedEx), admin dashboard, and notification services.",
    "version": "1.0.0",
    "contact": {
      "email": "brewed@lwangblack.co",
      "url": "https://www.lwangblack.co"
    }
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local development" },
    { "url": "https://api.lwangblack.co", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/logistics/carriers": {
      "get": {
        "tags": ["Logistics"],
        "summary": "List supported carriers",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Carrier list (DHL, FedEx, UPS, Shippo, Sajilo, Aramex)" } }
      }
    },
    "/api/logistics/zones": {
      "get": {
        "tags": ["Logistics"],
        "summary": "List delivery zones with pricing",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Delivery zones (Kathmandu free, global paid)" } }
      }
    },
    "/api/logistics/rates": {
      "post": {
        "tags": ["Logistics"],
        "summary": "Get shipping rates (Shippo API with zone fallback)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "toCountry": { "type": "string" },
                  "toCity": { "type": "string" },
                  "toZip": { "type": "string" },
                  "weight": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Rate quotes from carriers" } }
      }
    },
    "/api/logistics/create-shipment": {
      "post": {
        "tags": ["Logistics"],
        "summary": "Create shipment and generate tracking (Shippo label purchase)",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Tracking number, label URL, notifications sent" } }
      }
    },
    "/api/logistics/track": {
      "post": {
        "tags": ["Logistics"],
        "summary": "Track a shipment",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["trackingNumber"],
                "properties": {
                  "trackingNumber": { "type": "string" },
                  "carrierId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Tracking events and status" } }
      }
    },
...
    "/api/logistics/admin-override": {
      "post": {
        "tags": ["Logistics"],
        "summary": "Admin override shipping cost for an order",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Shipping cost overridden" } }
      }
    },
...
  }
}